Drop old test codes
