+++ tests/c2c.rs
@@ -1,72 +1,57 @@
 extern crate fftw;
-extern crate ndarray;
-#[macro_use]
-extern crate ndarray_linalg;
 extern crate num_traits;
 
-use fftw::*;
-use ndarray::*;
-use ndarray_linalg::*;
+use fftw::plan::*;
+use fftw::types::*;
+use num_traits::Zero;
 
 /// Check successive forward and backward transform equals to the identity
-fn test_identity<C: FFTWComplex>(mut pair: Pair<C, C, Ix1>, rtol: C::Real) {
-    let a: Array1<C> = random(pair.a.dim());
-    println!("a = {:?}", &a);
-    let b = pair.forward_array(a.view()).to_owned();
-    println!("b = {:?}", &b);
-    let a2 = pair.backward_array(b.view());
-    println!("a2 = {:?}", &a2);
-    assert_close_l2!(&a2, &a, rtol);
+#[test]
+fn c2c2c_identity() {
+    let n = 32;
+    let mut a = vec![c64::zero(); n];
+    let mut b = vec![c64::zero(); n];
+    let mut plan: C2CPlan64 =
+        C2CPlan::new(&[n], &mut a, &mut b, Sign::Forward, Flag::Measure).unwrap();
+    for i in 0..n {
+        a[i] = c64::new(1.0, 0.0);
+    }
+    plan.c2c(&mut a, &mut b).unwrap();
+    plan.c2c(&mut b, &mut a).unwrap();
+    for v in a.iter() {
+        let ans = c64::new(n as f64, 0.0);
+        let dif = (v - ans).norm();
+        if dif > 1e-7 {
+            panic!("Large difference: v={}, dif={}", v, dif);
+        }
+    }
 }
 
-/// Check successive forward and backward transform equals to the identity
-fn test_forward<C: FFTWComplex>(mut pair: Pair<C, C, Ix1>, rtol: C::Real) {
-    let n = pair.a.dim();
+/// Check cos transform
+#[test]
+fn c2c_cos() {
+    let n = 32;
+    let mut a = vec![c64::zero(); n];
+    let mut b = vec![c64::zero(); n];
+    let mut plan: C2CPlan64 =
+        C2CPlan::new(&[n], &mut a, &mut b, Sign::Forward, Flag::Measure).unwrap();
     let pi = ::std::f64::consts::PI;
-    let a: Array1<C> =
-        Array::from_iter((0..n).map(|i| Scalar::from_f64((2.0 * pi * i as f64 / n as f64).cos())));
-    println!("a = {:?}", &a);
-    let b = pair.forward_array(a.view()).to_owned();
-    println!("b = {:?}", &b);
-    // cos(x) = (exp(ix) + exp(-ix))/2
-    let mut ans: Array1<C> = Array::zeros(b.len());
-    ans[1] = Scalar::from_f64(0.5);
-    ans[n - 1] = Scalar::from_f64(0.5);
-    assert_close_l2!(&b, &ans, rtol);
-}
-
-mod c2c_64 {
-    use super::*;
-    const N: usize = 32;
-    const RTOL: f64 = 1e-7;
-
-    #[test]
-    fn identity() {
-        let pair: Pair<c64, c64, Ix1> = c2c_1d(N).to_pair().unwrap();
-        test_identity(pair, RTOL);
+    for i in 0..n {
+        a[i] = c64::new((2.0 * pi * i as f64 / n as f64).cos(), 0.0);
     }
-
-    #[test]
-    fn forward() {
-        let pair: Pair<c64, c64, Ix1> = c2c_1d(N).to_pair().unwrap();
-        test_forward(pair, RTOL);
-    }
-}
-
-mod c2c_32 {
-    use super::*;
-    const N: usize = 32;
-    const RTOL: f32 = 1e-4;
-
-    #[test]
-    fn identity() {
-        let pair: Pair<c32, c32, Ix1> = c2c_1d(N).to_pair().unwrap();
-        test_identity(pair, RTOL);
-    }
-
-    #[test]
-    fn forward() {
-        let pair: Pair<c32, c32, Ix1> = c2c_1d(N).to_pair().unwrap();
-        test_forward(pair, RTOL);
+    plan.c2c(&mut a, &mut b).unwrap();
+    for (i, v) in b.iter().enumerate() {
+        let ans = if i == 1 || i == n - 1 {
+            0.5 * n as f64
+        } else {
+            0.0
+        };
+        let dif = (v - ans).norm();
+        if dif > 1e-7 {
+            panic!(
+                "Large difference: v={}, ans={}, dif={}, i={}",
+                v, ans, dif, i
+            );
+        }
     }
 }
